fix build system for ppx_deriving.4.3
The build system is depending on the ppx_deriving.api package, which
is the right thing for ppx_deriving plugins, except that the
ppx_deriving_crowbar_runtime module also requires
a ppx_deriving.runtime (or ppx_deriving) dependency. This mistake in
the build system of the current package was previously unobservable as
ppx_deriving would install the object files of ppx_deriving.api and
ppx_deriving.runtime in the same directory, but this will not be the
case anymore with ppx_deriving.4.3.